+++ src/Guilded/Events/TopicReactionEvent.cs
@@ -10,18 +10,18 @@ using Newtonsoft.Json;
 namespace Guilded.Events;
 
 /// <summary>
-/// Represents an event that occurs when someone adds or removes <see cref="Content.Reaction">a reaction</see>.
+/// Represents an event that occurs when someone adds or removes a <see cref="Reaction">reaction</see>.
 /// </summary>
 /// <seealso cref="ReactionEvent{T}" />
 /// <seealso cref="TopicReactionEvent" />
 /// <seealso cref="Reaction" />
 /// <seealso cref="MessageEvent" />
 /// <seealso cref="MessageDeletedEvent" />
-public class TopicReactionEvent : ReactionEvent<MessageReaction>
+public class TopicReactionEvent : ReactionEvent<TopicReaction>
 {
     #region Properties
-    /// <inheritdoc cref="MessageReaction.MessageId" />
-    public Guid MessageId => Reaction.MessageId;
+    /// <inheritdoc cref="TopicReaction.TopicId" />
+    public uint TopicId => Reaction.TopicId;
     #endregion
 
     #region Constructors
@@ -35,7 +35,7 @@ public class TopicReactionEvent : ReactionEvent<MessageReaction>
     [JsonConstructor]
     public TopicReactionEvent(
         [JsonProperty(Required = Required.Always)]
-        MessageReaction reaction,
+        TopicReaction reaction,
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
         HashId? serverId